Uploaded Open Canvas
Open Canvas contains an outline of the entire project. It has the problem statement, how we would solve it, what would be our resources to solve it, what expertise do we need, who would be the target audience, how do we get contributions, and what activities do we plan for the audience/participants.